ファイル分割を実施
record-viewsが長いが、チューニングを通じて短くなると思うので、これ以上の分割は行わない
