+++ development/backend/src/api/files.js
@@ -1,0 +1,55 @@
+const fs = require('fs');
+const { v4: uuidv4 } = require('uuid');
+
+const jimp = require('jimp');
+
+const db = require("../mysql");
+
+// POST files/
+// ファイルのアップロード
+const postFiles = async (req, res) => {
+  let user = await db.getLinkedUser(req.headers);
+
+  if (!user) {
+    res.status(401).send();
+    return;
+  }
+
+  const base64Data = req.body.data;
+  mylog(base64Data);
+
+  const name = req.body.name;
+
+  const newId = uuidv4();
+  const newThumbId = uuidv4();
+
+  const binary = Buffer.from(base64Data, 'base64');
+
+  fs.writeFileSync(`${filePath}${newId}_${name}`, binary);
+
+  const image = await jimp.read(fs.readFileSync(`${filePath}${newId}_${name}`));
+  mylog(image.bitmap.width);
+  mylog(image.bitmap.height);
+
+  const size = image.bitmap.width < image.bitmap.height ? image.bitmap.width : image.bitmap.height;
+  await image.cover(size, size);
+
+  await image.writeAsync(`${filePath}${newThumbId}_thumb_${name}`);
+
+  await db.pool.query(
+    `insert into file (file_id, path, name)
+        values (?, ?, ?)`,
+    [`${newId}`, `${filePath}${newId}_${name}`, `${name}`],
+  );
+  await db.pool.query(
+    `insert into file (file_id, path, name)
+        values (?, ?, ?)`,
+    [`${newThumbId}`, `${filePath}${newThumbId}_thumb_${name}`, `thumb_${name}`],
+  );
+
+  res.send({ fileId: newId, thumbFileId: newThumbId });
+};
+
+module.exports = {
+  postFiles,
+};
