+++ development/backend/src/api/records-files.js
@@ -1,0 +1,90 @@
+const db = require("../mysql");
+
+// GET records/{recordId}/files/{itemId}
+// 添付ファイルのダウンロード
+const getRecordItemFile = async (req, res) => {
+  let user = await db.getLinkedUser(req.headers);
+
+  if (!user) {
+    res.status(401).send();
+    return;
+  }
+
+  const recordId = req.params.recordId;
+  mylog(recordId);
+  const itemId = Number(req.params.itemId);
+  mylog(itemId);
+
+  const [rows] = await db.pool.query(
+    `select f.name, f.path from record_item_file r
+    inner join file f
+    on
+    r.linked_record_id = ?
+    and
+    r.item_id = ?
+    and
+    r.linked_file_id = f.file_id`,
+    [`${recordId}`, `${itemId}`],
+  );
+
+  if (rows.length !== 1) {
+    res.status(404).send({});
+    return;
+  }
+  mylog(rows[0]);
+
+  const fileInfo = rows[0];
+
+  const data = fs.readFileSync(fileInfo.path);
+  const base64 = data.toString('base64');
+  mylog(base64);
+
+  res.send({ data: base64, name: fileInfo.name });
+};
+
+// GET records/{recordId}/files/{itemId}/thumbnail
+// 添付ファイルのサムネイルダウンロード
+const getRecordItemFileThumbnail = async (req, res) => {
+  let user = await db.getLinkedUser(req.headers);
+
+  if (!user) {
+    res.status(401).send();
+    return;
+  }
+
+  const recordId = req.params.recordId;
+  mylog(recordId);
+  const itemId = Number(req.params.itemId);
+  mylog(itemId);
+
+  const [rows] = await db.pool.query(
+    `select f.name, f.path from record_item_file r
+    inner join file f
+    on
+    r.linked_record_id = ?
+    and
+    r.item_id = ?
+    and
+    r.linked_thumbnail_file_id = f.file_id`,
+    [`${recordId}`, `${itemId}`],
+  );
+
+  if (rows.length !== 1) {
+    res.status(404).send({});
+    return;
+  }
+  mylog(rows[0]);
+
+  const fileInfo = rows[0];
+
+  const data = fs.readFileSync(fileInfo.path);
+  const base64 = data.toString('base64');
+  mylog(base64);
+
+  res.send({ data: base64, name: fileInfo.name });
+};
+
+module.exports = {
+  getRecordItemFile,
+  getRecordItemFileThumbnail,
+};
