+++ development/backend/src/index.js
@@ -4,6 +4,12 @@ const app = express();
 app.use(express.json({limit: '10mb'}))
 
 const api = require("./api");
+const api_categories = require("./api/categories");
+const api_files = require("./api/files");
+const api_record_views = require("./api/record-views");
+const api_records_comments = require("./api/records-comments");
+const api_records_files = require("./api/records-files");
+const api_records = require("./api/records");
 
 app.get('/api/hello', (req, res) => {
   console.log('requested');
@@ -12,7 +18,7 @@ app.get('/api/hello', (req, res) => {
 
 app.post('/api/client/records', async (req, res, next) => {
   try {
-    await api.postRecords(req, res);
+    await api_records.postRecords(req, res);
   } catch(e) {
     console.log(e);
     next(new Error("Unexpect"));
@@ -21,7 +27,7 @@ app.post('/api/client/records', async (req, res, next) => {
 
 app.get('/api/client/records/:recordId', async (req, res, next) => {
   try {
-    await api.getRecord(req, res);
+    await api_records.getRecord(req, res);
   } catch(e) {
     console.log(e);
     next(new Error("Unexpect"));
@@ -30,7 +36,7 @@ app.get('/api/client/records/:recordId', async (req, res, next) => {
 
 app.get('/api/client/record-views/tomeActive', async (req, res, next) => {
   try {
-    await api.tomeActive(req, res);
+    await api_record_views.tomeActive(req, res);
   } catch(e) {
     console.log(e);
     next(new Error("Unexpect"));
@@ -39,7 +45,7 @@ app.get('/api/client/record-views/tomeActive', async (req, res, next) => {
 
 app.get('/api/client/record-views/allActive', async (req, res, next) => {
   try {
-    await api.allActive(req, res);
+    await api_record_views.allActive(req, res);
   } catch(e) {
     console.log(e);
     next(new Error("Unexpect"));
@@ -48,7 +54,7 @@ app.get('/api/client/record-views/allActive', async (req, res, next) => {
 
 app.get('/api/client/record-views/allClosed', async (req, res, next) => {
   try {
-    await api.allClosed(req, res);
+    await api_record_views.allClosed(req, res);
   } catch(e) {
     console.log(e);
     next(new Error("Unexpect"));
@@ -57,7 +63,7 @@ app.get('/api/client/record-views/allClosed', async (req, res, next) => {
 
 app.get('/api/client/record-views/mineActive', async (req, res, next) => {
   try {
-    await api.mineActive(req, res);
+    await api_record_views.mineActive(req, res);
   } catch(e) {
     console.log(e);
     next(new Error("Unexpect"));
@@ -66,7 +72,7 @@ app.get('/api/client/record-views/mineActive', async (req, res, next) => {
 
 app.put('/api/client/records/:recordId', async (req, res, next) => {
   try {
-    await api.updateRecord(req, res);
+    await api_records.updateRecord(req, res);
   } catch(e) {
     console.log(e);
     next(new Error("Unexpect"));
@@ -75,7 +81,7 @@ app.put('/api/client/records/:recordId', async (req, res, next) => {
 
 app.get('/api/client/records/:recordId/comments', async (req, res, next) => {
   try {
-    await api.getComments(req, res);
+    await api_records_comments.getComments(req, res);
   } catch(e) {
     console.log(e);
     next(new Error("Unexpect"));
@@ -84,7 +90,7 @@ app.get('/api/client/records/:recordId/comments', async (req, res, next) => {
 
 app.post('/api/client/records/:recordId/comments', async (req, res, next) => {
   try {
-    await api.postComments(req, res);
+    await api_records_comments.postComments(req, res);
   } catch(e) {
     console.log(e);
     next(new Error("Unexpect"));
@@ -93,7 +99,7 @@ app.post('/api/client/records/:recordId/comments', async (req, res, next) => {
 
 app.get('/api/client/categories', async (req, res, next) => {
   try {
-    await api.getCategories(req, res);
+    await api_categories.getCategories(req, res);
   } catch(e) {
     console.log(e);
     next(new Error("Unexpect"));
@@ -102,7 +108,7 @@ app.get('/api/client/categories', async (req, res, next) => {
 
 app.post('/api/client/files', async (req, res, next) => {
   try {
-    await api.postFiles(req, res);
+    await api_files.postFiles(req, res);
   } catch(e) {
     console.log(e);
     next(new Error("Unexpect"));
@@ -111,7 +117,7 @@ app.post('/api/client/files', async (req, res, next) => {
 
 app.get('/api/client/records/:recordId/files/:itemId', async (req, res, next) => {
   try {
-    await api.getRecordItemFile(req, res);
+    await api_records_files.getRecordItemFile(req, res);
   } catch(e) {
     console.log(e);
     next(new Error("Unexpect"));
@@ -120,7 +126,7 @@ app.get('/api/client/records/:recordId/files/:itemId', async (req, res, next) =>
 
 app.get('/api/client/records/:recordId/files/:itemId/thumbnail', async (req, res, next) => {
   try {
-    await api.getRecordItemFileThumbnail(req, res);
+    await api_records_files.getRecordItemFileThumbnail(req, res);
   } catch(e) {
     console.log(e);
     next(new Error("Unexpect"));
