+++ development/backend/src/mysql.js
@@ -1,0 +1,46 @@
+const mysql = require('mysql2/promise');
+
+// MEMO: 設定項目はここを参考にした
+// https://github.com/sidorares/node-mysql2#api-and-configuration
+// https://github.com/mysqljs/mysql
+const mysqlOption = {
+  host: 'mysql',
+  user: 'backend',
+  password: 'backend',
+  database: 'app',
+  waitForConnections: true,
+  connectionLimit: 10,
+};
+const pool = mysql.createPool(mysqlOption);
+
+const mylog = (obj) => {
+  if (Array.isArray(obj)) {
+    for (const e of obj) {
+      console.log(e);
+    }
+    return;
+  }
+  console.log(obj);
+};
+
+const getLinkedUser = async (headers) => {
+  const target = headers['x-app-key'];
+  mylog(target);
+  const qs = `select * from session where value = ?`;
+
+  const [rows] = await pool.query(qs, [`${target}`]);
+
+  if (rows.length !== 1) {
+    mylog('セッションが見つかりませんでした。');
+    return undefined;
+  }
+
+  return { user_id: rows[0].linked_user_id };
+};
+
+const filePath = 'file/';
+
+module.exports = {
+  pool,
+  getLinkedUser,
+};
